Implement memset to zero
Implement memset to zero function and call it
in performExtCodeCopy and OP_CALLDATACOPY. Main
loop of this implementation is done in 3
instructions (store with increment, compare and
conditional jump) compared to the old implementation
that was done 6 instructions. For processing rest
bytes, calls to mstore8 are removed and changed to
do in a single store.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>