fix: fix forge permissions + L2GenesisUpgrade test (#1848)
# What ❔

The permissions for forge broke. Seems like the "read" access overwrites
"read-write" access control that comes AFTER it.
Thus, all the "read-write"-s are moved to the beginning of the list.

Also, in the `L2GenesisUpgrade.t.sol` there was a mock missing found
while debugging
[this](https://github.com/matter-labs/foundry-zksync/issues/1210). This
PR adds it.
<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.